src/backend/requirements.txt
alembic==1.17.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==5.0.0
black==25.9.0
build==1.3.0
CacheControl==0.14.3
certifi==2025.10.5
cfgv==3.4.0
charset-normalizer==3.4.4
cleo==2.1.0
click==8.3.0
colorama==0.4.6
coverage==7.11.2
crashtest==0.4.1
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
dulwich==0.24.10
email-validator==2.3.0
fastapi==0.121.1
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
filelock==3.20.0
findpython==0.7.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
inflect==7.5.0
iniconfig==2.3.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
Jinja2==3.1.6
jiter==0.12.0
keyring==25.6.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
openai==2.7.2
orjson==3.11.4
packaging==25.0
pathspec==0.12.1
pbs-installer==2025.10.31
pkginfo==1.12.1.2
platformdirs==4.5.0
pluggy==1.6.0
poetry==2.2.1
poetry-core==2.2.1
pre_commit==4.4.0
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
ruff==0.14.4
sentry-sdk==2.43.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tomlkit==0.13.3
tqdm==4.67.1
trove-classifiers==2025.9.11.17
typeguard==4.4.4
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
zstandard==0.25.0
